move heading tag component out of shadcn section
